fix: meter attribute getter arg types

+++ src/native/main.cpp
@@ -246,11 +246,11 @@ Napi::Value getHostLicenseMeterAttribute(const Napi::CallbackInfo &info)
     }
     STRING arg0 = toEncodedString(info[0].As<Napi::String>());
     Napi::BigInt64Array array1 = info[1].As<Napi::BigInt64Array>();
-    uint32_t *arg1 = reinterpret_cast<int64_t *>(array1.ArrayBuffer().Data());
+    int64_t *arg1 = reinterpret_cast<int64_t *>(array1.ArrayBuffer().Data());
     Napi::BigInt64Array array2 = info[2].As<Napi::BigInt64Array>();
-    uint32_t *arg2 = reinterpret_cast<uint64_t *>(array2.ArrayBuffer().Data());
+    uint64_t *arg2 = reinterpret_cast<uint64_t *>(array2.ArrayBuffer().Data());
     Napi::BigInt64Array array3 = info[3].As<Napi::BigInt64Array>();
-    uint32_t *arg3 = reinterpret_cast<uint64_t *>(array3.ArrayBuffer().Data());
+    uint64_t *arg3 = reinterpret_cast<uint64_t *>(array3.ArrayBuffer().Data());
     return Napi::Number::New(env, GetHostLicenseMeterAttribute(arg0.c_str(), arg1, arg2, arg3));
 }
